+++ rnalysis/filtering.py
@@ -4437,8 +4437,8 @@ class CountFilter(Filter):
     @readable_name('Principal Component Analysis (PCA) plot')
     def pca(self, samples: Union[param_typing.GroupedColumns, Literal['all']] = 'all', n_components: PositiveInt = 3,
             power_transform: bool = True, labels: bool = True, title: Union[str, Literal['auto']] = 'auto',
-            title_fontsize: float = 20, label_fontsize: float = 16, tick_fontsize: float = 12
-            ) -> Tuple[PCA, List[plt.Figure]]:
+            title_fontsize: float = 20, label_fontsize: float = 16, tick_fontsize: float = 12,
+            proportional_axes: bool = False) -> Tuple[PCA, List[plt.Figure]]:
         """
         Performs Principal Component Analysis (PCA), visualizing the principal components that explain the most\
         variance between the different samples. The function will standardize the data prior to PCA, and then plot \
@@ -4466,6 +4466,9 @@ class CountFilter(Filter):
         :type label_fontsize: float (default=15)
         :param tick_fontsize: determines the font size of the X and Y tick labels, and the sample name labels. .
         :type tick_fontsize: float (default=10)
+        :param proportional_axes: if True, the dimensions of the PCA plots will be proportional \
+        to the percentage of variance explained by each principal component.
+        :type proportional_axes: bool (default=False)
         :return: A tuple whose first element is an sklearn.decomposition.pca object, \
         and second element is a list of matplotlib.axis objects.
 
@@ -4502,7 +4505,7 @@ class CountFilter(Filter):
                         [f'Principal component {1 + first_pc}', f'Principal component {1 + second_pc}', 'lib']],
                     pc1_var=pc_var[first_pc], pc2_var=pc_var[second_pc], sample_grouping=samples, labels=labels,
                     label_fontsize=label_fontsize, title=title, title_fontsize=title_fontsize,
-                    tick_fontsize=tick_fontsize))
+                    tick_fontsize=tick_fontsize, proportional_axes=proportional_axes))
 
         return pca_obj, figs
 
@@ -4527,7 +4530,8 @@ class CountFilter(Filter):
 
     @staticmethod
     def _pca_plot(final_df: pd.DataFrame, pc1_var: float, pc2_var: float, sample_grouping: param_typing.GroupedColumns,
-                  labels: bool, title: str, title_fontsize: float, label_fontsize: float, tick_fontsize: float):
+                  labels: bool, title: str, title_fontsize: float, label_fontsize: float, tick_fontsize: float,
+                  proportional_axes: bool):
         """
         Internal method, used to plot the results from CountFilter.pca().
 
@@ -4542,8 +4546,15 @@ class CountFilter(Filter):
         :return: an axis object containing the PCA plot.
 
         """
-        fig = plt.figure(figsize=(8, 8))
+        if proportional_axes:
+            dims = (15, 15 * (pc2_var / pc1_var))
+        else:
+            dims = (8, 8)
+        fig = plt.figure(figsize=dims)
         ax = fig.add_subplot(1, 1, 1)
+        if proportional_axes:
+            ax.set_aspect(pc2_var / pc1_var)
+
         ax.grid(True)
         ax.set_xlabel(f'{final_df.columns[0]} (explained {pc1_var * 100 :.2f}%)', fontsize=label_fontsize)
         ax.set_ylabel(f'{final_df.columns[1]} (explained {pc2_var * 100 :.2f}%)', fontsize=label_fontsize)